+++ db/migrate/20260906000001_add_label_name_to_things.rb
@@ -1,0 +1,5 @@
+class AddLabelNameToThings < ActiveRecord::Migration[8.1]
+  def change
+    add_column :things, :label_name, :string
+  end
+end